# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Systock API",
        "description": "Systock Challenge API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/produtos": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos",
                "operationId": "cab8aec265fc5e5ab4a1d09399ead918",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos retornada com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Criar novo produto",
                "operationId": "848c58aee26092da8f35ffed07a79969",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "preco",
                                    "descricao"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Notebook Gamer"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4500
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Notebook potente para jogos e edição"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/produtos/{id}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Buscar produto por ID",
                "operationId": "298e77ccdf34792d1e380c79685ec28c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto encontrado"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Atualizar produto",
                "operationId": "dabb26799e32a04fdb911e59080a6696",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "preco": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "descricao": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Remover produto",
                "operationId": "32c920c8c4df5eb965b1bd0a63c021fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto removido com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/produtos/usuario/{usuarioId}": {
            "get": {
                "tags": [
                    "Produtos"
                ],
                "summary": "Listar produtos de um usuário específico",
                "operationId": "40ac638128d78a706b4b0d27d17e8018",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de produtos filtrada por usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/relatorio": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Relatório Geral (Usuários e Médias)",
                "operationId": "db7e7d7c5f4dcd7148f6da5755e5b224",
                "responses": {
                    "200": {
                        "description": "Relatório geral retornado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/relatorio/maiores-estoques": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Usuários com mais produtos",
                "operationId": "7dfc14e0de8f2d3f410eec1595e2635b",
                "responses": {
                    "200": {
                        "description": "Ranking de estoques retornado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/relatorio/produtos-mais-caros": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Produto mais caro de cada usuário",
                "operationId": "fab6f86fbc445fa404bd87c64da351ab",
                "responses": {
                    "200": {
                        "description": "Produtos premium por usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/relatorio/faixas-precos": {
            "get": {
                "tags": [
                    "Relatórios"
                ],
                "summary": "Quantidade de produtos por faixa de preço",
                "operationId": "6d4fc10a07a58df63d71f5cd6b5cdf38",
                "responses": {
                    "200": {
                        "description": "Análise de precificação retornada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}